- Rev 2999 GOGOGO  - More game constants  - Fixed annoying bug of Admins/GMs getting into the GM map and losing buffs (?)  - Line ending fixes O.o :D  - I hope this one goes well
git-svn-id: svn://70.113.44.214/VanaMirror/trunk@2999 09c1253f-3d09-8e40-9719-4320e56601b5

+++ src/CashServer/SyncPacket.h
@@ -28,6 +28,6 @@ class WorldServerConnection;
 namespace SyncPacket {
 	void playerChangeChannel(WorldServerConnection *player, Player *info, uint16_t channel);
 	void playerBuffsTransferred(WorldServerConnection *player, int32_t playerid);
-	void registerPlayer(WorldServerConnection *player, uint32_t ip, int32_t playerid, const string &name, int32_t map, int32_t job, int32_t level, int32_t guildid, uint8_t guildrank, int32_t allianceid, uint8_t alliancerank);
+	void registerPlayer(WorldServerConnection *player, uint32_t ip, int32_t playerid, const string &name, int32_t map, int16_t job, uint8_t level, int32_t guildid, uint8_t guildrank, int32_t allianceid, uint8_t alliancerank);
 	void removePlayer(WorldServerConnection *player, int32_t playerid);
 }
